requirements.txt
kafka-python
pyspark
vaderSentiment
psycopg2-binary
SQLAlchemy
streamlit
pandas
requests
beautifulsoup4
lxml